+++ claasp/ciphers/block_ciphers/aradi_block_cipher_sbox_and_compact_linear_map.py
@@ -72,7 +72,7 @@ class AradiBlockCipherSBoxAndCompactLinearMap(Cipher):
 
     INPUT:
 
-    - ``block_bit_size`` -- **integer** (default: `64`); cipher input and output block bit size of the cipher
+    - ``block_bit_size`` -- **integer** (default: `64`); the bit size of the cipher's input and output blocks.
     - ``key_bit_size`` -- **integer** (default: `128`); cipher key bit size of the cipher
     - ``number_of_rounds`` -- **integer** (default: `0`); number of rounds of the cipher. The cipher uses the
       corresponding amount given the other parameters (if available) when number_of_rounds is 0
